Robot/Light functionality change
The robot behavior used to be randomized, now their values are hardcoded
to make them avoid light.

The lighting used to be more colorful, now it is reduced to tiny circles
of a point of light. But this light does not spread further than a well
defined circle. It makes the robots behavior more obvious.